requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.3
amqp==5.1.1
amqplib==1.0.2
anyjson==0.3.3
appdirs==1.4.4
APScheduler==3.9.1.post1
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==19.10b0
blinker==1.5
Brotli==1.0.9
bs4==0.0.1
by==0.0.4
cached-property==1.5.2
carrot==0.10.7
celery==5.2.7
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coloredlogs==15.0.1
configparser==5.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
ddddocr==1.4.7
Deprecated==1.2.13
distlib==0.3.6
Django==3.2.10
django-apscheduler==0.6.2
django-bootstrap3==21.2
django-cors-headers==3.13.0
django-filter==22.1
django-picklefield==3.1
django-redis==5.2.0
django-smart-selects==1.6.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
dnspython==2.3.0
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
drf-yasg==1.21.4
et-xmlfile==1.1.0
eventlet==0.33.3
exceptiongroup==1.0.1
execnet==1.9.0
Faker==17.0.0
filelock==3.9.0
Flask==2.0.2
flatbuffers==22.9.24
form==0.0.1
future==0.18.2
gitdb2==2.0.6
GitPython==2.1.15
greenlet==2.0.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httprunner==3.1.8
humanfriendly==10.0
hyperframe==6.0.1
idna==2.7
importlib-metadata==4.13.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.9.5
Js2Py==0.74
jsonpath==0.82
jsonschema==2.6.0
kaitaistruct==0.10
kombu==5.2.4
logging3==0.0.2
loguru==0.4.1
MarkupSafe==2.0.1
mpmath==1.2.1
mypy-extensions==0.4.3
nose==1.3.7
numpy==1.21.6
onnxruntime==1.12.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
pathspec==0.10.1
pefile==2022.5.30
pick==2.0.2
pika==1.3.1
Pillow==9.0.1
platformdirs==2.5.3
pluggy==1.0.0
poco==0.98.1
prompt-toolkit==3.0.36
protobuf==4.21.7
psycopg2==2.9.3
py==1.11.0
pyaml==18.11.0
pyasn1==0.4.8
pycparser==2.21
pycryptodomex==3.15.0
pydantic==1.8.2
pydivert==2.1.0
PyExecJS==1.5.1
PyGithub==1.43.5
Pygments==2.12.0
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.13
pyjsparser==2.7.1
PyJWT==1.7.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
pypiwin32==223
Pypubsub==4.0.3
pyreadline==2.1
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.2.0
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist==2.4.0
pytest-yaml==1.2.1
python-alipay-sdk==3.1.2
python-dateutil==2.8.2
python-gitlab==1.8.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==302
pywin32-ctypes==0.2.0
PyYAML==5.4.1
redis==4.5.1
regex==2022.10.31
requests==2.27.1
robotframework==5.0
robotframework-pythonlibcore==3.0.0
robotframework-ride==1.7.4.2
robotframework-selenium2library==3.0.0
robotframework-seleniumlibrary==6.0.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==3.7.0
sentry-sdk==0.14.4
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
sniffio==1.2.0
some-package==0.1
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
svn==0.3.46
sympy==1.10.1
texttable==0.9.1
toml==0.10.2
tomli==2.0.1
tqdm==4.63.1
trio==0.19.0
trio-websocket==0.9.2
typed-ast==1.5.4
typing-extensions==3.10.0.2
tzdata==2022.6
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.19.0
wcwidth==0.2.6
websocket-client==0.59.0
Werkzeug==2.0.2
win32-setctime==1.1.0
windows-curses==2.3.0
wrapt==1.13.3
wsproto==1.0.0
wxPython==4.0.7.post2
xlrd==1.2.0
xlutils3==1.0.2
xlwings==0.27.10
xlwt==1.3.0
zipp==3.6.0
zstandard==0.19.0
